Added Player script
I've started working on methods involved in the Player script. I didn't
test them yet, but I wrote methods to calculate the player's rank and
some movement related methods. Those movement related methods should be
able to move the player one space at a time (so it can stop at
junctions/star spaces) and snap the player 30 units above its space on
the z-axis when it isn't their turn (will need to adjust this one so it
can account for multiple players on the same space).