did a bunch of work on gamecontroller and player class
the game now technically works.  players roll for initative, perform a
turn, and start a minigame after the turn is over.  It will pick a
minigame based on the spaces the players land on.  Upgraded the movement
method in player to move to any arbitrary point.  Added a method to run
code once when a player state is changed to make code cleaner and less
redundant