gameBoard data is saved when transitioning between scenes
FINALLY.  This solution is pretty bad and requires us to manually
designate which variables we want to save.  Essentially some objects
look to a dontdestroyonload object for their parameters, and we save the
latest data before a new scene is loaded.  So when the game board is
loaded again, they can take the data from the gamedata object