Revert "Added "hats", adjusted camera angle"
This reverts commit 3d6f57d20973291e3e4969b058e07ec76ed16928.

Some fire came up when I tried to commit and sync a bit ago. The git
shell told me some fatal error ocurred with a meta file for the
handcart2 prefab. I deleted it since it was the meta file, I hope that
didn't break anything too much.

This is pretty much the revert to the revert I accidentally made while
trying to work out how to handle the git shell.